DDL scanner works with class Foo extends Table[..]. also created some tests
