+++ src/main/resources/play.plugins
@@ -1,1 +1,1 @@
-400:play.api.db.slick.SlickDDLPlugin
+400:play.api.db.slick.plugin.SlickDDLPlugin
